% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{print.banner}
\alias{print.banner}
\title{Print method for banner objects}
\usage{
\method{print}{banner}(x, ...)
}
\arguments{
\item{x}{A character string vector as procuced by banner()}

\item{...}{Not used}
}
\value{
\code{x} itself, invisibly. A side effect is that \code{x}
is transferred to a \code{clipboard} device, if possible
}
\description{
As well as printing the comment string in the console window
the same text strings are transferred to a clipboard, if possible,
ready for pasting into the R script currently being drafted.
}
